# PERF: DataFrame reductions with axis=None and EA dtypes